refactor builtin
