+++ alphabot/builtins/commands/basic/index.js
@@ -1,0 +1,7 @@
+const command = require('./command');
+const vanish = require('./vanish');
+
+module.exports = {
+  command,
+  vanish,
+}
