+++ alphabot/builtins/commands/moderation/commercial.js
@@ -1,0 +1,17 @@
+const client = require('../../../src/alphabot');
+const logger = require('../../../config/logger');
+const {isNumber, isFinite} = require('../../../utils/numbers');
+
+const commercial = async (channel, [duration = 30, filler]) => {
+  if(filler !== undefined || !isNumber(duration) && !isFinite(duration)) return;
+
+  client.commercial(channel, duration)
+    .then((data) => {
+      logger.info(channel + ' is now running an commercial for  ' + duration + ' seconds! | ' + data);
+    }).catch((err) => {
+      logger.error(err);
+    }
+  );
+};
+
+module.exports = commercial;
