+++ alphabot/chat/response/responseHandler.js
