+++ alphabot/chat/response/responseParser.js
