+++ alphabot/db/commandFunctions.js
@@ -93,9 +93,9 @@ const findAllCommands = async (user) => {
       commands.push(v);
     });
     if (commands.length === 0) {
-      logger.info('User has no commands');
+      logger.info('User has no response');
     } else {
-      logger.info('Found commands');
+      logger.info('Found response');
     }
   });
